Add routing with react router, navigation component, shop component
