+++ src/routes/shop/shop.component.jsx
@@ -1,0 +1,5 @@
+const Shop = () => {
+  return <h1>I am the shop page</h1>;
+};
+
+export default Shop;
